requirements.txt
# ============================================================
# ConnectorOS Scout — Python Dependencies
# Install: pip install -r requirements.txt
# ============================================================

# --- Core ---
python-dotenv==1.0.0
pydantic==2.6.0
structlog==24.1.0

# --- Database ---
sqlalchemy==2.0.25
alembic==1.13.0

# --- API Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0

# --- HTTP Client ---
httpx==0.26.0
tenacity==8.2.3

# --- AI ---
openai==1.12.0

# --- Job Board APIs ---
google-search-results==2.4.2

# --- Excel Export ---
openpyxl==3.1.2
psycopg2-binary==2.9.9

# --- Data Processing ---
pandas==2.2.0

# --- DNS Verification ---
dnspython==2.5.0

# --- Fuzzy Matching ---
thefuzz==0.22.1

# --- Scheduling ---
apscheduler==3.10.4

# --- YAML Config ---
pyyaml==6.0.1
